+++ packages/webview_flutter/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewFlutterPlugin.java
@@ -4,8 +4,11 @@
 
 package io.flutter.plugins.webviewflutter;
 
+import android.app.Activity;
 import android.content.Context;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
 
 import com.tencent.smtt.export.external.TbsCoreSettings;
 import com.tencent.smtt.sdk.QbSdk;
@@ -13,7 +16,12 @@ import com.tencent.smtt.sdk.QbSdk;
 import java.util.HashMap;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.common.EventChannel;
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
 
 /**
  * Java platform implementation of the webview_flutter plugin.
@@ -23,7 +31,12 @@ import io.flutter.plugin.common.BinaryMessenger;
  * <p>Call {@link #registerWith(Registrar)} to use the stable {@code io.flutter.plugin.common}
  * package instead.
  */
-public class WebViewFlutterPlugin implements FlutterPlugin {
+public class WebViewFlutterPlugin implements FlutterPlugin, ActivityAware,
+        MethodChannel.MethodCallHandler {
+    private static final String CHANNEL_NAME = "webview_flutter_x5";
+    private static Activity mActivity;
+    private static WebViewFactory webViewFactory;
+    private MethodChannel channel;
 
     private FlutterCookieManager flutterCookieManager;
 
@@ -51,15 +64,49 @@ public class WebViewFlutterPlugin implements FlutterPlugin {
      */
     @SuppressWarnings("deprecation")
     public static void registerWith(io.flutter.plugin.common.PluginRegistry.Registrar registrar) {
+        mActivity = registrar.activity();
+        webViewFactory = new WebViewFactory(registrar.messenger(),
+                registrar.view(), mActivity);
+        registrar.addActivityResultListener(webViewFactory);
         registrar
                 .platformViewRegistry()
                 .registerViewFactory(
                         "plugins.flutter.io/webview",
-                        new WebViewFactory(registrar.messenger(), registrar.view()));
+                        webViewFactory);
         new FlutterCookieManager(registrar.messenger());
+        final WebViewFlutterPlugin plugin = new WebViewFlutterPlugin();
+        plugin.setupChannel(registrar.messenger(), registrar.context().getApplicationContext());
+    }
+
+    @Override
+    public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+        mActivity = binding.getActivity();
+        if (webViewFactory != null) {
+            webViewFactory.setActivity(mActivity);
+        }
+        binding.addActivityResultListener(webViewFactory);
+    }
+
+    @Override
+    public void onDetachedFromActivityForConfigChanges() {
+        mActivity = null;
+    }
+
+    @Override
+    public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+        mActivity = binding.getActivity();
+        if (webViewFactory != null) {
+            webViewFactory.setActivity(mActivity);
+        }
+    }
+
+    @Override
+    public void onDetachedFromActivity() {
+        mActivity = null;
     }
 
     void initX5(Context context) {
+        Log.i("X5_webView", "准备初始化");
 //        TBS内核首次使用和加载时，ART虚拟机会将Dex文件转为Oat，该过程由系统底层触发且耗时较长，很容易引起anr问题，解决方法是使用TBS的 ”dex2oat优化方案“。
 // 在调用TBS初始化、创建WebView之前进行如下配置
         HashMap map = new HashMap();
@@ -71,12 +118,12 @@ public class WebViewFlutterPlugin implements FlutterPlugin {
             @Override
             public void onViewInitFinished(boolean arg0) {
                 //x5內核初始化完成的回调，为true表示x5内核加载成功，否则表示x5内核加载失败，会自动切换到系统内核。
-                Log.i("X5 webView", "onViewInitFinished:" + arg0);
+                Log.i("X5_webView", "onViewInitFinished:" + arg0);
             }
 
             @Override
             public void onCoreInitFinished() {
-                Log.i("X5 webView", "onCoreInitFinished");
+                Log.i("X5_webView", "onCoreInitFinished");
             }
         };
         //x5内核初始化接口
@@ -86,21 +133,45 @@ public class WebViewFlutterPlugin implements FlutterPlugin {
     @Override
     public void onAttachedToEngine(FlutterPluginBinding binding) {
         BinaryMessenger messenger = binding.getBinaryMessenger();
+        webViewFactory = new WebViewFactory(messenger,
+                /*containerView=*/ null, mActivity);
         binding
                 .getPlatformViewRegistry()
                 .registerViewFactory(
-                        "plugins.flutter.io/webview", new WebViewFactory(messenger, /*containerView=*/ null));
+                        "plugins.flutter.io/webview", webViewFactory);
         flutterCookieManager = new FlutterCookieManager(messenger);
-        initX5(binding.getApplicationContext());
+        setupChannel(binding.getBinaryMessenger(),
+                binding.getApplicationContext());
     }
 
     @Override
     public void onDetachedFromEngine(FlutterPluginBinding binding) {
+        channel.setMethodCallHandler(null);
+        channel = null;
+        mActivity = null;
         if (flutterCookieManager == null) {
             return;
         }
 
         flutterCookieManager.dispose();
         flutterCookieManager = null;
+    }
+
+    private void setupChannel(BinaryMessenger messenger, Context context) {
+        channel = new MethodChannel(messenger, CHANNEL_NAME);
+        channel.setMethodCallHandler(this);
+    }
+
+    @Override
+    public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
+        switch (call.method) {
+            case "initX5":
+                initX5(mActivity.getApplicationContext());
+                result.success("");
+                break;
+            default:
+                result.notImplemented();
+                break;
+        }
     }
 }